Refactor Footer style and move header container to component
