+++ common/actions/header.js
@@ -1,5 +1,10 @@
 export const MENU_CLICK = 'MENU_CLICK';
+export const RESIZE = 'RESIZE';
 
 export const menuClick = () => ({
   type: MENU_CLICK,
+});
+
+export const resize = () => ({
+  type: RESIZE,
 });
