+++ common/reducers/header.js
@@ -1,9 +1,11 @@
-import { MENU_CLICK } from '../actions/header';
+import { MENU_CLICK, RESIZE } from '../actions/header';
 
 const header = (state = false, action) => {
   switch (action.type) {
     case MENU_CLICK:
       return !state;
+    case RESIZE:
+      return false;
     default:
       return state;
   }
